move component into a dedicated file [skip ci pytest]
